+++ shared/src/commonMain/kotlin/hibernate/v2/model/transport/RouteDetailsStop.kt
@@ -2,5 +2,5 @@ package hibernate.v2.model.transport
 
 data class RouteDetailsStop(
     val transportStop: TransportStop,
-    var savedEtaId: Long?
+    var savedEtaId: Int?
 )
